# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: URL Shortener API
  description: |
    A comprehensive API for URL shortening with analytics, user management, subscriptions, and payment processing.

    ## Authentication
    Most endpoints require authentication via JWT Bearer token or HTTP-only cookies.
    - **Access Token**: Short-lived token for API authentication
    - **Refresh Token**: Long-lived token for obtaining new access tokens

    ## Rate Limiting
    API requests may be rate limited based on your subscription plan.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:5080/api/v1
    description: Development server
  - url: https://api.example.com/api/v1
    description: Production server
...
paths:
...
  /auth/refresh-token:
    post:
      tags:
        - Auth
      summary: Refresh access token
      description: Get a new access token using a refresh token from request body or cookies.
      operationId: refreshToken
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-email:
    post:
      tags:
        - Auth
      summary: Verify email address
      description: Verify user email using the token sent via email.
      operationId: verifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequest"
      responses:
        "200":
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/resend-verification:
    post:
      tags:
        - Auth
      summary: Resend verification email
      description: Resend the email verification link to the user's email address.
      operationId: resendVerificationEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendVerificationRequest"
      responses:
        "200":
          description: Verification email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      tags:
        - Auth
      summary: Request password reset
      description: Send a password reset link to the user's email address.
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Password reset link sent to email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
  /auth/reset-password:
    post:
      tags:
        - Auth
      summary: Reset password
      description: Reset user password using the token received via email.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me/avatar:
    post:
      tags:
        - Auth
      summary: Upload avatar
      description: Upload a new avatar image for the current user.
      operationId: uploadAvatar
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
                  description: Avatar image file (JPEG, PNG, GIF)
              required:
                - avatar
      responses:
        "200":
          description: Avatar uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Avatar uploaded successfully
                  data:
                    type: object
                    properties:
                      avatarUrl:
                        type: string
                        format: uri
                        example: https://example.com/avatars/user123.jpg
        "400":
          description: Invalid file type or size
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/change-password:
    put:
      tags:
        - Auth
      summary: Change password
      description: Change the current user's password.
      operationId: changePassword
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          description: Current password is incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/devices:
    get:
      tags:
        - Auth
      summary: Get logged in devices
      description: Get a list of all devices where the user is currently logged in.
      operationId: getDevices
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Devices list retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevicesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/devices/{id}:
    delete:
      tags:
        - Auth
      summary: Logout from specific device
      description: Invalidate a specific device session by its ID.
      operationId: logoutDevice
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "200":
          description: Device logged out successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Device not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==================== URL ROUTES ====================
...
  /clicks/analytics:
    get:
      tags:
        - Analytics
      summary: Get user analytics
      description: Get overall analytics for all user's URLs.
      operationId: getMyAnalytics
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: urlId
          in: query
          description: Filter analytics by specific URL ID
          schema:
            type: string
        - name: days
          in: query
          description: Number of days to include in analytics
          schema:
            type: integer
            minimum: 1
            default: 30
      responses:
        "200":
          description: Analytics retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /clicks/analytics/{urlId}:
    get:
      tags:
        - Analytics
      summary: Get URL analytics
      description: Get detailed analytics for a specific URL.
      operationId: getUrlAnalytics
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: urlId
          in: path
          required: true
          description: URL ID
          schema:
            type: string
        - name: days
          in: query
          description: Number of days to include in analytics
          schema:
            type: integer
            minimum: 1
            default: 30
      responses:
        "200":
          description: URL analytics retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UrlAnalyticsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /clicks/url/{urlId}:
    get:
      tags:
        - Analytics
      summary: Get URL clicks
      description: Get paginated list of clicks for a specific URL.
      operationId: getUrlClicks
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: urlId
          in: path
          required: true
          description: URL ID
          schema:
            type: string
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: startDate
          in: query
          description: Filter clicks from this date (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: Filter clicks until this date (ISO 8601)
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Clicks retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClicksListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ==================== PLANS ROUTES ====================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: HTTP-only cookie containing JWT access token

  # ==================== PARAMETERS ====================
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    DeviceId:
      name: id
      in: path
      required: true
      description: Device/session ID
      schema:
        type: string

  # ==================== RESPONSES ====================
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            message: Authentication required
            statusCode: 401
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            message: Resource not found
            statusCode: 404
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"

  # ==================== SCHEMAS ====================
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
      required:
        - success
        - message
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: An error occurred
        statusCode:
          type: integer
          example: 400
      required:
        - success
        - message
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: email
              message:
                type: string
                example: Invalid email format
      required:
        - success
        - message
        - errors
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10
        hasNextPage:
          type: boolean
          example: true
        hasPrevPage:
          type: boolean
          example: false

    # Auth Schemas
...
    RefreshTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
          description: Refresh token (optional if using cookies)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    VerifyEmailRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 1
          example: verification-token-here
      required:
        - token
    ResendVerificationRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
      required:
        - email
    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: john.doe@example.com
      required:
        - email
    ResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 1
          example: reset-token-here
        password:
          type: string
          minLength: 6
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)"
          example: NewPassword123
      required:
        - token
        - password
...
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
          minLength: 1
          example: OldPassword123
        newPassword:
          type: string
          minLength: 6
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)"
          example: NewPassword456
      required:
        - currentPassword
        - newPassword
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Login successful
        data:
          type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
            accessToken:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            refreshToken:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...

    # User Schemas
    User:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        id:
          type: string
          example: 507f1f77bcf86cd799439011
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        role:
          type: string
          enum: [admin, user]
          example: user
        avatar:
          type: string
          nullable: true
          example: https://example.com/avatar.jpg
        isActive:
          type: boolean
          example: true
        isEmailVerified:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
    DevicesResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                example: session-id-123
              device:
                type: string
                example: Chrome on Windows
              ip:
                type: string
                example: 192.168.1.1
              lastActive:
                type: string
                format: date-time
                example: "2024-01-15T10:30:00Z"
              isCurrent:
                type: boolean
                example: true

    # URL Schemas
    Url:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        id:
          type: string
          example: 507f1f77bcf86cd799439011
        originalUrl:
          type: string
          format: uri
          example: https://example.com/very/long/url/path
        shortId:
          type: string
          example: abc123
        shortUrl:
          type: string
          format: uri
          example: https://short.url/abc123
        user:
          type: string
          example: 507f1f77bcf86cd799439011
        clickCount:
          type: integer
          example: 150
        isActive:
          type: boolean
          example: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
          example: "2024-12-31T23:59:59Z"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
    Click:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        url:
          type: string
          example: 507f1f77bcf86cd799439011
        ip:
          type: string
          example: 192.168.1.1
        userAgent:
          type: string
          example: Mozilla/5.0 (Windows NT 10.0; Win64; x64)...
        referer:
          type: string
          nullable: true
          example: https://google.com
        device:
          type: object
          properties:
            type:
              type: string
              example: desktop
            browser:
              type: string
              example: Chrome
            os:
              type: string
              example: Windows
        location:
          type: object
          properties:
            country:
              type: string
              example: United States
            city:
              type: string
              example: New York
            region:
              type: string
              example: NY
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    ClicksListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Click"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    AnalyticsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            totalClicks:
              type: integer
              example: 5000
            uniqueVisitors:
              type: integer
              example: 3500
            clicksByDay:
              type: array
              items:
                type: object
                properties:
                  date:
                    type: string
                    format: date
                    example: "2024-01-15"
                  clicks:
                    type: integer
                    example: 150
            topCountries:
              type: array
              items:
                type: object
                properties:
                  country:
                    type: string
                    example: United States
                  clicks:
                    type: integer
                    example: 2000
            topDevices:
              type: array
              items:
                type: object
                properties:
                  device:
                    type: string
                    example: desktop
                  clicks:
                    type: integer
                    example: 3000
            topBrowsers:
              type: array
              items:
                type: object
                properties:
                  browser:
                    type: string
                    example: Chrome
                  clicks:
                    type: integer
                    example: 2500
    UrlAnalyticsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            url:
              $ref: "#/components/schemas/Url"
            analytics:
              type: object
              properties:
                totalClicks:
                  type: integer
                  example: 500
                uniqueVisitors:
                  type: integer
                  example: 350
                clicksByDay:
                  type: array
                  items:
                    type: object
                    properties:
                      date:
                        type: string
                        format: date
                        example: "2024-01-15"
                      clicks:
                        type: integer
                        example: 25
                topCountries:
                  type: array
                  items:
                    type: object
                    properties:
                      country:
                        type: string
                        example: United States
                      clicks:
                        type: integer
                        example: 200
                topDevices:
                  type: array
                  items:
                    type: object
                    properties:
                      device:
                        type: string
                        example: desktop
                      clicks:
                        type: integer
                        example: 300
                topBrowsers:
                  type: array
                  items:
                    type: object
                    properties:
                      browser:
                        type: string
                        example: Chrome
                      clicks:
                        type: integer
                        example: 250
                topReferers:
                  type: array
                  items:
                    type: object
                    properties:
                      referer:
                        type: string
                        example: https://google.com
                      clicks:
                        type: integer
                        example: 100

    # Plan Schemas
...